+++ pages/[lang]/pure-motivation-and-great-conviction-2.tsx
@@ -1,0 +1,10 @@
+import Moti from "components/Motivation";
+import React, { FC } from "react";
+
+const Motivation: FC<{}> = () => {
+  /* Render */
+  // return <The />;
+  return <Moti />;
+};
+
+export default Motivation;
